### 2022-01-24 #4097 @whedon
[PRE REVIEW]: pyemgpipeline: A Python package for electromyography processing
**Submitting author:** @tlwu (<a href="http://orcid.org/0000-0002-1823-6818">Tsung-Lin Wu</a>)
**Repository:** <a href="https://github.com/aalhossary/pyemgpipeline" target ="_blank">https://github.com/aalhossary/pyemgpipeline</a>
**Version:** v0.1.0
