### 2022-02-03 #4097 @tuliofalmeida
Hi,

I appreciate the invitation. It will be a pleasure to review this paper.

Best,

Túlio
From: Øystein Sørensen ***@***.***>
Sent: quinta-feira, 3 de fevereiro de 2022 02:50
Cc: Tulio De Almeida ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: pyemgpipeline: A Python package for electromyography processing (Issue #4097)

@tuliofalmeida<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Ftuliofalmeida&data=04%7C01%7C%7C738487d075db4838ce6008d9e6d90b3c%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C637794642152430419%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000&sdata=3RcKN9D%2FA2nZTvi%2BuCX3SCkeBfErNPejZync%2FchFKdI%3D&reserved=0> would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fjoss.readthedocs.io%2Fen%2Flatest%2Freview_criteria.html&data=04%7C01%7C%7C738487d075db4838ce6008d9e6d90b3c%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C637794642152430419%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000&sdata=pNv8mJPajs1Gsn1XT0h2IY1bMf3rLd%2BylSloRhOaaqA%3D&reserved=0>

-
Triage notifications on the go with GitHub Mobile for iOS<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fapps.apple.com%2Fapp%2Fapple-store%2Fid1477376905%3Fct%3Dnotification-email%26mt%3D8%26pt%3D524675&data=04%7C01%7C%7C738487d075db4838ce6008d9e6d90b3c%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C637794642152430419%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000&sdata=lsF%2BVHnH8HMB8BdofQB3D1NR8CDRW5HWZZujg9cBArE%3D&reserved=0> or Android<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fplay.google.com%2Fstore%2Fapps%2Fdetails%3Fid%3Dcom.github.android%26referrer%3Dutm_campaign%253Dnotification-email%2526utm_medium%253Demail%2526utm_source%253Dgithub&data=04%7C01%7C%7C738487d075db4838ce6008d9e6d90b3c%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C637794642152430419%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000&sdata=RL7eYeHbj2pOLvY%2BBpMo%2Fnt1jvTWbH%2Bta%2FEyThrTR6o%3D&reserved=0>.
You are receiving this because you were mentioned.Message ID: ***@***.******@***.***>>
