### 2022-02-12 #4156 @whedon
[REVIEW]: pyemgpipeline: A Python package for electromyography processing
**Submitting author:** @tlwu (<a href="http://orcid.org/0000-0002-1823-6818">Tsung-Lin Wu</a>)
**Repository:** https://github.com/aalhossary/pyemgpipeline
**Branch with paper.md** (empty if default branch):
**Version:** version 1.0.0_RC
