### 2022-02-20 aalhossary/pyemgpipeline#1 @tuliofalmeida
Review JOSS - Clinician
This is an issue created as part of [pyemgpipeline JOSS Review](https://github.com/openjournals/joss-reviews/issues/4156)

Comments :

The library is well documented, easy to install and meets an existing demand.
I believe that the examples can be improved, especially considering the clinician with little experience in Python, this being my main point of observation.

Suggestions:

- Create a more interactive example/tutorial, using google colab (use the markdowns to explain the functions). The example in google colab can be run all with a single command.

- Perform a library test script (just join the example code in a script).

- Develop an example with some dataset available online (maybe this one: https://zenodo.org/record/5239861, or even more interesting if linked to some publication.)

- Document and better explain the data input, this manipulation will be the most difficult for the user with little experience in Python.

- The plot parameters function allows customizing the plot. For example: change color, dpi, labels, overplot ?
