+++ paper/paper.bib
@@ -4,7 +4,7 @@
     journal = {Journal of Electromyography and Kinesiology},
     year = {2016},
     volume = {31},
-    pages = {126–135},
+    pages = {126-135},
     doi = {10.1016/j.jelekin.2016.10.001},
 }
 
@@ -14,7 +14,7 @@
     journal = {Journal of Electromyography and Kinesiology},
     year = {2006},
     volume = {16},
-    pages = {175–187},
+    pages = {175-187},
     doi = {10.1016/j.jelekin.2005.07.003},
 }
 
@@ -51,7 +51,7 @@
 
 @techreport{Stegeman:1999,
     author = {Stegeman, D.F. and Hermens, H.J.},
-    title = {Standards for surface electromyography: the European project "Surface EMG for non invasive assessment of muscles (SENIAM)"},
+    title = {Standards for surface electromyography: the European project "Surface EMG for non-invasive assessment of muscles (SENIAM)"},
     institution = {European Community},
     year = {1999},
     note = {European concerted action, Biomed 2 program},
