### 2022-03-28 0672880
Re-organize iterating utility
