+++ pyemgpipeline/utilities/__init__.py
@@ -1,0 +1,5 @@
+from . utility import iter_dict_or_list
+
+__all__ = [
+    "iter_dict_or_list",
+]
