### 2022-03-28 817f18f
Expand visualization (line2d param, trials overlapping)
