+++ pyemgpipeline/plots/__init__.py
@@ -1,7 +1,9 @@
 from . emg_plot import EMGPlotParams
 from . emg_plot import plot_emg
+from . emg_plot import plot_emg_overlapping_trials
 
 __all__ = [
     "EMGPlotParams",
     "plot_emg",
+    "plot_emg_overlapping_trials",
 ]
