### 2022-03-28 ec2e070
Make all unit tests run altogether without using IDE
