+++ tests/test_processors/__init__.py
@@ -1,0 +1,19 @@
+from . test_amplitude_normalizer import TestAmplitudeNormalizer
+from . test_bandpass_filter import TestBandpassFilter
+from . test_base_processor import TestBaseProcessor
+from . test_dc_offset_remover import TestDCOffsetRemover
+from . test_end_frame_cutter import TestEndFrameCutter
+from . test_full_wave_rectifier import TestFullWaveRectifier
+from . test_linear_envelope import TestLinearEnvelope
+from . test_segmenter import TestSegmenter
+
+__all__ = [
+    'TestAmplitudeNormalizer',
+    'TestBandpassFilter',
+    'TestBaseProcessor',
+    'TestDCOffsetRemover',
+    'TestEndFrameCutter',
+    'TestFullWaveRectifier',
+    'TestLinearEnvelope',
+    'TestSegmenter',
+]
