+++ tests/__init__.py
@@ -1,0 +1,2 @@
+from . test_processors import *
+from . test_wrappers import *
