+++ tests/test_wrappers/__init__.py
@@ -1,0 +1,9 @@
+from . test_data_processing_manager import TestDataProcessingManager
+from . test_emg_measurement import TestEMGMeasurement
+from . test_emg_measurement_collection import  TestEMGMeasurementCollection
+
+__all__ = [
+    'TestDataProcessingManager',
+    'TestEMGMeasurement',
+    'TestEMGMeasurementCollection',
+]
