### 2022-03-30 c998dea
Refactor wrapper classes (eg not allow multiple trials in dict)
