+++ pyemgpipeline/wrappers/emg_measurement_collection.py
@@ -1,6 +1,5 @@
 from .. processors import *
 from .. plots import *
-from .. utilities import iter_dict_or_list
 from . emg_measurement import EMGMeasurement
 import numpy as np
 import copy
@@ -11,10 +10,8 @@ class EMGMeasurementCollection:
 
     Parameters
     ----------
-    all_data : dict or list
-        If dict, keys can be trial names and values are signal data of
-        the trials.
-        If list, elements are signal data of the trials.
+    all_data : list
+        Elements of all_data are signal data of the trials.
         Signal data of each trial should be ndarray of shape
         (n_samples,) or (n_samples, n_channels), where n_samples > 15
         (See Notes).
@@ -31,11 +28,7 @@ class EMGMeasurementCollection:
         For fine wire EMG, the authors suggest a minimum sample rate of
         2000 Hz.
 
-    all_timestamp : dict, list, or None, default None
-        If dict or list, all_timestamp should be of the same type as
-        all_data.
-        If dict, keys should be identical to those of all_data and
-        values are ndarray or None.
+    all_timestamp : list or None, default None
         If list, its length should be the same as the length of all_data
         and elements are ndarray or None.
         The ndarray in all_timestamp is the actual timestamp of the
@@ -43,17 +36,15 @@ class EMGMeasurementCollection:
         length as the first dimension of its corresponding element in
         all_data.
 
+    trial_names : list or None, default None
+        Trial names.
+        If not None, trial_names and all_data should have the same
+        length.
+
     channel_names : list or None, default None
         If list, elements are str and its length should be equal to
         n_channels.
         Channel names of all trials to be shown in plots.
-
-    all_main_titles : list or None, default None
-        The main title in the plot, which is valid only when all_data
-        is a list. (If all_data is a dict, its keys will be used as
-        main titles.)
-        If not None, all_main_titles and all_data should have the same
-        length.
 
     emg_plot_params : EMGPlotParams or None, default None
         See class EMGPlotParams and function emg_plot.
@@ -77,16 +68,16 @@ class EMGMeasurementCollection:
         action. 108-112.
     """
 
-    def __init__(self, all_data, hz, all_timestamp=None, channel_names=None,
-                 all_main_titles=None, emg_plot_params=None):
-        assert isinstance(all_data, dict) or isinstance(all_data, list), 'all_data must be dict or list'
-        for k in iter_dict_or_list(all_data):
+    def __init__(self, all_data, hz, all_timestamp=None,
+                 trial_names=None, channel_names=None, emg_plot_params=None):
+        assert isinstance(all_data, list), 'all_data must be list'
+        for k in range(len(all_data)):
             BaseProcessor.assert_input(all_data[k])
             assert all_data[k].shape[0] > 15, 'first dimension of ndarray in all_data[k] must have length > 15'
-        ndims = [all_data[k].ndim for k in iter_dict_or_list(all_data)]
+        ndims = [all_data[k].ndim for k in range(len(all_data))]
         assert all([e == ndims[0] for e in ndims]), 'Dimensions of all trials should be identical'
         if ndims[0] == 2:
-            nchs = [all_data[k].shape[1] for k in iter_dict_or_list(all_data)]
+            nchs = [all_data[k].shape[1] for k in range(len(all_data))]
             assert all([e == nchs[0] for e in nchs]), 'n_channels of all trials should be identical'
         self.all_data = copy.deepcopy(all_data)
 
@@ -94,35 +85,22 @@ class EMGMeasurementCollection:
 
         self.all_timestamp = copy.deepcopy(all_timestamp)
         if self.all_timestamp is not None:
[…]
