+++ tests/test_wrappers/test_emg_measurement_collection.py
@@ -6,49 +6,20 @@ from pyemgpipeline.wrappers import EMGMeasurementCollection
 class TestEMGMeasurementCollection(TestCase):
     def setUp(self):
         hz = 1000
-
-        # case: dict
-        all_data_dict = {
-            'T1': np.array([20.3, 41.0, 53.9, 63.3, 39.5, 24.9, 26.1, 24.0, 44.1, 42.0,
-                            37.4, 24.6, -21.8, -56.3, -48.1, -45.0, -29.1, -9.6, 5.3, 1.4]),
-            'T2': np.array([1.1, 2.9, 1.4, -0.2, 4.4, 7.2, 9.9, 19.1, 14.2, 18.8,
-                            17.2, 17.9, 11.1, 13.9, 15.4, 19.4, 12.1, 16.9, 12.4, 9.0])
-        }
-        self.c_in_dict = EMGMeasurementCollection(all_data_dict, hz)
-
-        # case: list
-        all_data_list = [
+        all_data = [
             np.array([20.3, 41.0, 53.9, 63.3, 39.5, 24.9, 26.1, 24.0, 44.1, 42.0,
                       37.4, 24.6, -21.8, -56.3, -48.1, -45.0, -29.1, -9.6, 5.3, 1.4]),
             np.array([1.1, 2.9, 1.4, -0.2, 4.4, 7.2, 9.9, 19.1, 14.2, 18.8,
                       17.2, 17.9, 11.1, 13.9, 15.4, 19.4, 12.1, 16.9, 12.4, 9.0])
         ]
-        self.c_in_list = EMGMeasurementCollection(all_data_list, hz)
+        self.c = EMGMeasurementCollection(all_data, hz)
 
     def tearDown(self):
         pass
 
     def test_apply_dc_offset_remover(self):
-        # case: dict
-        self.c_in_dict.apply_dc_offset_remover()
-        actual = self.c_in_dict.all_data
-        desired = {
-            'T1': np.array([8.405, 29.105, 42.005, 51.405, 27.605,
-                            13.005, 14.205, 12.105, 32.205, 30.105,
-                            25.505, 12.705, -33.695, -68.195, -59.995,
-                            -56.895, -40.995, -21.495, -6.595, -10.495]),
-            'T2': np.array([-10.105, -8.305, -9.805, -11.405, -6.805,
-                            -4.005, -1.305, 7.895, 2.995, 7.595,
-                            5.995, 6.695, -0.105, 2.695, 4.195,
-                            8.195, 0.895, 5.695, 1.195, -2.205])
-        }
-        self.assertEqual(actual.keys(), desired.keys())
-        for k in actual.keys():
-            np.testing.assert_allclose(actual[k], desired[k])
-
-        # case: list
-        self.c_in_list.apply_dc_offset_remover()
-        actual = self.c_in_list.all_data
+        self.c.apply_dc_offset_remover()
+        actual = self.c.all_data
         desired = [
             np.array([8.405, 29.105, 42.005, 51.405, 27.605,
                       13.005, 14.205, 12.105, 32.205, 30.105,
@@ -68,28 +39,9 @@ class TestEMGMeasurementCollection(TestCase):
         bf_cutoff_fq_lo = 10
         bf_cutoff_fq_hi = 450
 
-        # case: dict
-        self.c_in_dict.apply_bandpass_filter(
+        self.c.apply_bandpass_filter(
             bf_order=bf_order, bf_cutoff_fq_lo=bf_cutoff_fq_lo, bf_cutoff_fq_hi=bf_cutoff_fq_hi)
-        actual = self.c_in_dict.all_data
-        desired = {
-            'T1': np.array([-12.9626930, 9.3485823, 23.9736743, 35.1195602, 12.6483173,
[…]
