+++ setup.py
@@ -4,8 +4,8 @@ with open('README.md', 'r', encoding='utf-8') as fh:
     long_description = fh.read()
 
 setup(
-    name='pyemgpipeline',
-    version='0.1.0',
+    name='pyemgpipeline-20220331',
+    version='0.1.1',
     author='tlwu et al',
     author_email='tlwu2008@gmail.com',
     description='EMG signal processing pipeline',
