+++ pyemgpipeline/utilities/__init__.py
@@ -1,5 +1,0 @@
-from . utility import iter_dict_or_list
-
-__all__ = [
-    "iter_dict_or_list",
-]

[…]
