### 2022-03-31 897b24d
Add more params to manage legend display
