+++ pyemgpipeline/wrappers/data_processing_manager.py
@@ -261,9 +261,9 @@ class DataProcessingManager:
         if self.segmenter is not None:
             print(f'Segmenter            : {self.segmenter.get_param_values_in_str()}')
 
-    def _plot_processing_chain(self, current_step,
-                               k_for_plot=None, is_overlapping_trials=False, cycled_colors=None,
-                               emg_plot_params=None):
+    def _plot_processing_chain(self, current_step, k_for_plot=None, emg_plot_params=None,
+                               is_overlapping_trials=False,
+                               cycled_colors=None, legend_kwargs=None, axes_pos_adjust=None):
         """Plot current step of the processing chain
 
         Parameters
@@ -278,15 +278,34 @@ class DataProcessingManager:
             is_overlapping_trials is False, only trial k_for_plot will
             be plotted.
 
+        emg_plot_params : EMGPlotParams or None, default None
+            See class EMGPlotParams and function emg_plot.
+
         is_overlapping_trials : bool, default False
             Whether overlapping trials of the same channel.
             If True, all trials will be plotted and overlapped.
 
         cycled_colors : list or None, default None
+            cycled_colors is used when is_overlapping_trials is True.
             The colors for plotting overlapped trials data.
 
-        emg_plot_params : EMGPlotParams or None, default None
-            See class EMGPlotParams and function emg_plot.
+        legend_kwargs : dict or None, default None
+            legend_kwargs is used when is_overlapping_trials is True.
+            Parameters to control the legend display. They are the
+            "other parameters" of method matplotlib.axes.Axes.legend,
+            including loc, bbox_to_anchor, ncol, prop, fontsize, etc.
+            (See
+            https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.legend.html).
+
+        axes_pos_adjust : 4-tuple or None, default None
+            axes_pos_adjust is used when is_overlapping_trials is True.
+            Parameters to adjust the axes position (i.e., plot position)
+            when legend is displayed to prevent legend overlaying the plot.
+            The 4-tuple represents: [0] shift of left position relative to
+            width, [1] shift of bottom position relative to height, [2]
+            proportion of width, [3] proportion of height.
+            If None, no adjustment is applied, i.e., the value (0, 0, 1, 1)
+            is applied.
 
         Returns
         -------
@@ -294,9 +313,9 @@ class DataProcessingManager:
         """
         if is_overlapping_trials:
             plot_emg_overlapping_trials(self.c.all_data, self.c.all_timestamp, self.c.trial_names,
-                                        cycled_colors=cycled_colors,
                                         channel_names=self.c.channel_names, main_title=current_step,
-                                        emg_plot_params=emg_plot_params)
+                                        emg_plot_params=emg_plot_params, cycled_colors=cycled_colors,
+                                        legend_kwargs=legend_kwargs, axes_pos_adjust=axes_pos_adjust)
         else:
             for k in range(len(self.c.all_data)):
                 if k_for_plot is not None and k != k_for_plot:
@@ -307,7 +326,8 @@ class DataProcessingManager:
                          emg_plot_params=emg_plot_params)
 
     def process_all(self, is_plot_processing_chain=False, k_for_plot=None,
-                    is_overlapping_trials=False, cycled_colors=None):
+                    is_overlapping_trials=False,
+                    cycled_colors=None, legend_kwargs=None, axes_pos_adjust=None):
         """Apply current processors to data and plot intermediate results
 
         Parameters
@@ -328,7 +348,26 @@ class DataProcessingManager:
             If True, all trials will be plotted and overlapped.
 
         cycled_colors : list or None, default None
[…]
