+++ setup.py
@@ -4,7 +4,7 @@ with open('README.md', 'r', encoding='utf-8') as fh:
     long_description = fh.read()
 
 setup(
-    name='pyemgpipeline-20220331',
+    name='pyemgpipeline',
     version='0.1.1',
     author='tlwu et al',
     author_email='tlwu2008@gmail.com',
