### 2022-04-01 aalhossary/pyemgpipeline#1 @tlwu
@tuliofalmeida Thank you for the suggestions on several points to make this library more user friendly. We have revised it based on the suggestions.

>     * Create a more interactive example/tutorial, using google colab (use the markdowns to explain the functions). The example in google colab can be run all with a single command.

The [examples](https://aalhossary.github.io/pyemgpipeline/examples.html) in the documentation are revised such that the old and new ones can all be opened in google colab with explanations and links to api references. Now the users can easily play around the notebooks interactively on colab.

>     * Perform a library test script (just join the example code in a script).

The unittest scripts are now organized and can be executed with a single command, i.e., "python -m unittest discover -v". More details are put [here](https://github.com/aalhossary/pyemgpipeline/blob/master/tests/readme__run_all_unittest).

>     * Develop an example with some dataset available online (maybe this one: https://zenodo.org/record/5239861, or even more interesting if linked to some publication.)

Examples 1~4 in [examples](https://aalhossary.github.io/pyemgpipeline/examples.html) use datasets available online as sample data to demonstrate this library. Datasets for example 1, 2, and 3 are from the [UC Irvine Machine Learning Repository](https://archive-beta.ics.uci.edu/ml/datasets). Newly added example 4 uses the knee OA dataset from Zenodo you suggested, which is an interesting case.

>     * Document and better explain the data input, this manipulation will be the most difficult for the user with little experience in Python.

A [notebook (i.e., example 0)](https://aalhossary.github.io/pyemgpipeline/notebooks/ex0_input_data_description.html) is added to describe the input data format in more details and discuss about converting raw data files into desired format.

>     * The plot parameters function allows customizing the plot. For example: change color, dpi, labels, overplot ?

Previously only dpi and labels are supported. Now changing color and overplotting are supported too. In summary,
-In class [EMGPlotParams](https://aalhossary.github.io/pyemgpipeline/api/pyemgpipeline.plots.html#emgplotparams), the parameter *fig_kwargs* controls the figure's size, dpi, spacing between subplots, etc.
-In class [EMGPlotParams](https://aalhossary.github.io/pyemgpipeline/api/pyemgpipeline.plots.html#emgplotparams), the parameter *line2d_kwargs* controls the color, properties of the lines and markers, etc.
-When multiple trials of the same channel are overlapping, the parameters of the method [*plot_emg_overlapping_trials*](https://aalhossary.github.io/pyemgpipeline/api/pyemgpipeline.plots.html#plot-emg-overlapping-trials) control the legend's labels and positions.
-In method [*plot_emg*](https://aalhossary.github.io/pyemgpipeline/api/pyemgpipeline.plots.html#plot-emg) and [*plot_emg_overlapping_trials*](https://aalhossary.github.io/pyemgpipeline/api/pyemgpipeline.plots.html#plot-emg-overlapping-trials), parameter *main_title* sets the main title while parameter *channel_names* sets the subplot titles.
