### 2022-04-04 aalhossary/pyemgpipeline#1 @tuliofalmeida
Nice! Congratulations on the library.
