### 2022-04-04 aalhossary/pyemgpipeline#2 @osorensen
Editorial comments to JOSS submission
I am the editor for the [JOSS submission of this software](https://github.com/openjournals/joss-reviews/issues/4156). I will be posting editorial comments for the manuscript as Issues in this repository, and this is the first one.

The following sentences in **Summary** do not seem 100 % complete:

> As in sport tasks particularly, it has been observed that amplitudes greater that 100% maximum voluntary contraction (MVC) can be observed [@Devaprakash:2016].  Therefore, we will be using amplitude normalization recommendations from @Devaprakash:2016 for amplitude normalization – this method guarantees a muscle will not exceed 100% MVC as all EMG trials from the experiment are used to identify maximal muscle activation.

I suggest rewriting to something like

> In sport tasks particularly, it has been observed that amplitudes greater that 100% maximum voluntary contraction (MVC) can be observed [@Devaprakash:2016]. Therefore, we will be using amplitude normalization recommendations from @Devaprakash:2016 for amplitude normalization – this method guarantees a muscle will not exceed 100% MVC as all EMG trials from the experiment are used to identify maximal muscle activation.
