### 2022-04-04 aalhossary/pyemgpipeline#3 @osorensen
Editorial comments to JOSS submission
Editorial comment to https://github.com/openjournals/joss-reviews/issues/4156.

In the following sentence in **Summary**

> What is not included in the package is time event detection and time normalization as different laboratories are interested in different phases of gait and prefer to use either linear or nonlinear time normalized techniques.

Should *normalized techniques* be changed to *normalization techniques*?
