### 2022-04-04 aalhossary/pyemgpipeline#4 @osorensen
Editorial comment to JOSS submission
Comment to https://github.com/openjournals/joss-reviews/issues/4156.

The following sentence in **Summary**:

> Researchers have thus tried to standardize the use of EMG regarding implementation, analysis, and reporting [@Stegeman:1999; @Merletti:1999; @JEK:2018], these fine works however have arguably been unsuccessfully adopted by many researchers to date.

I think it would make more sense to split it two, e.g.,

> Researchers have thus tried to standardize the use of EMG regarding implementation, analysis, and reporting [@Stegeman:1999; @Merletti:1999; @JEK:2018]. These fine works however have arguably been unsuccessfully adopted by many researchers to date.
