### 2022-04-07 e28cad7
Allow plotting selected trials
