+++ pyemgpipeline/wrappers/data_processing_manager.py
@@ -261,7 +261,7 @@ class DataProcessingManager:
         if self.segmenter is not None:
             print(f'Segmenter            : {self.segmenter.get_param_values_in_str()}')
 
-    def _plot_processing_chain(self, current_step, k_for_plot=None, emg_plot_params=None,
+    def _plot_processing_chain(self, current_step, trial_indices_for_plot, emg_plot_params=None,
                                is_overlapping_trials=False,
                                cycled_colors=None, legend_kwargs=None, axes_pos_adjust=None):
         """Plot current step of the processing chain
@@ -271,19 +271,16 @@ class DataProcessingManager:
         current_step : str
             Current step in the processing chain.
 
-        k_for_plot : index of list, or None
-            If k_for_plot is None, all trials will be plotted.
-            If k_for_plot is not None, it should be an integer between
-            0 and len(all_data) - 1. In this case, if
-            is_overlapping_trials is False, only trial k_for_plot will
-            be plotted.
+        trial_indices_for_plot : list of integer
+            A list of selected indices of the self.c.all_data,
+            indicating which trials of the data to be plotted.
 
         emg_plot_params : EMGPlotParams or None, default None
             See class EMGPlotParams and function emg_plot.
 
         is_overlapping_trials : bool, default False
-            Whether overlapping trials of the same channel.
-            If True, all trials will be plotted and overlapped.
+            Whether or not to plot trials of the same channel
+            overlappingly on one (sub)figure.
 
         cycled_colors : list or None, default None
             cycled_colors is used when is_overlapping_trials is True.
@@ -312,21 +309,21 @@ class DataProcessingManager:
         None
         """
         if is_overlapping_trials:
-            plot_emg_overlapping_trials(self.c.all_data, self.c.all_timestamp, self.c.trial_names,
+            legend_labels = [self.c.trial_names[k] for k in trial_indices_for_plot]
+            plot_emg_overlapping_trials(self.c.all_data, self.c.all_timestamp,
+                                        trial_indices_for_plot, legend_labels,
                                         channel_names=self.c.channel_names, main_title=current_step,
                                         emg_plot_params=emg_plot_params, cycled_colors=cycled_colors,
                                         legend_kwargs=legend_kwargs, axes_pos_adjust=axes_pos_adjust)
         else:
-            for k in range(len(self.c.all_data)):
-                if k_for_plot is not None and k != k_for_plot:
-                    continue
+            for k in trial_indices_for_plot:
                 plot_emg(self.c.all_data[k], self.c.all_timestamp[k],
                          channel_names=self.c.channel_names,
                          main_title=f'{current_step} ({self.c.trial_names[k]})',
                          emg_plot_params=emg_plot_params)
 
-    def process_all(self, is_plot_processing_chain=False, k_for_plot=None,
-                    is_overlapping_trials=False,
+    def process_all(self, is_plot_processing_chain=False,
+                    k_for_plot=None, is_overlapping_trials=False,
                     cycled_colors=None, legend_kwargs=None, axes_pos_adjust=None):
         """Apply current processors to data and plot intermediate results
 
@@ -336,16 +333,18 @@ class DataProcessingManager:
             Whether to plot the intermediate results after each
             processing step.
 
-        k_for_plot : index of list, or None
-            If k_for_plot is None, all trials will be plotted.
-            If k_for_plot is not None, it should be an integer between
-            0 and len(all_data) - 1. In this case, if
-            is_overlapping_trials is False, only trial k_for_plot will
[…]
