### 2022-04-07 16a26eb
Allow hiding legend
