+++ paper/paper.bib
@@ -36,6 +36,7 @@
     volume = {9},
     number = {1},
     pages = {III-IV},
+    doi = {10.1201/9781420036985.ax2},
 }
 
 @article{Pataky:2019,
