+++ paper/paper.md
@@ -33,9 +33,9 @@ bibliography: paper.bib
 
 # Summary
 
-We have developed an electromyography (EMG) signal processing pipeline package called `pyemgpipeline`, which is suitable for both surface EMG and intramuscular EMG processing.  `pyemgpipeline` implements internationally accepted EMG processing conventions and provides a high-level interface for ensuring user adherence to those conventions, in terms of (1) processing parameter values, (2) processing steps, and (3) processing step order.  The international standards are from surface EMG for non-invasive assessment of muscles (SENIAM) [@Stegeman:1999].  The seven processing steps included in the package are DC offset removal, bandpass filtering, full wave rectification, linear envelope, end frame cutting, amplitude normalization, and segmentation.  As in sport tasks particularly, it has been observed that amplitudes greater that 100% maximum voluntary contraction (MVC) can be observed [@Devaprakash:2016].  Therefore, we will be using amplitude normalization recommendations from @Devaprakash:2016 for amplitude normalization – this method guarantees a muscle will not exceed 100% MVC as all EMG trials from the experiment are used to identify maximal muscle activation.  What is not included in the package is time event detection and time normalization as different laboratories are interested in different phases of gait and prefer to use either linear or nonlinear time normalized techniques.
+We have developed an electromyography (EMG) signal processing pipeline package called `pyemgpipeline`, which is suitable for both surface EMG and intramuscular EMG processing.  `pyemgpipeline` implements internationally accepted EMG processing conventions and provides a high-level interface for ensuring user adherence to those conventions, in terms of (1) processing parameter values, (2) processing steps, and (3) processing step order.  The international standards are from surface EMG for non-invasive assessment of muscles (SENIAM) [@Stegeman:1999].  The seven processing steps included in the package are DC offset removal, bandpass filtering, full wave rectification, linear envelope, end frame cutting, amplitude normalization, and segmentation.  In sport tasks particularly, it has been observed that amplitudes greater that 100% maximum voluntary contraction (MVC) can be observed [@Devaprakash:2016].  Therefore, we will be using amplitude normalization recommendations from @Devaprakash:2016 for amplitude normalization – this method guarantees a muscle will not exceed 100% MVC as all EMG trials from the experiment are used to identify maximal muscle activation.  What is not included in the package is time event detection and time normalization as different laboratories are interested in different phases of gait and prefer to use either linear or nonlinear time normalization techniques.
 
-As stated by @Stegeman:1999 although EMG is easy to use it is also easy to abuse.  Researchers have thus tried to standardize the use of EMG regarding implementation, analysis, and reporting [@Stegeman:1999; @Merletti:1999; @JEK:2018], these fine works however have arguably been unsuccessfully adopted by many researchers to date.  Based on the published standards and decades of practical experience, we summarize EMG signal processing conventions into the seven steps mentioned above and implement them within an easy-to-use package so that researchers and clinicians with all levels of experience process their EMG signals using correct conventions.  In addition to the processing steps and their order, the choice of processing/filtering parameters can be changed to match the needs of the research or clinician.  For example, within the bandpass filtering step, proper values of the low and high cutoff frequencies of the Butterworth bandpass filter depend on different use cases.  We carefully set default values in the package, and we also include suggested values under different scenarios in [related API documentation](https://aalhossary.github.io/pyemgpipeline/api/pyemgpipeline.processors.html#bandpassfilter) [@JEK:2018; @Merletti:1999; @Stegeman:1999; @Drake:2006].
[…]
