+++ docs/_sources/examples.rst.txt
@@ -1,11 +1,0 @@
-Examples
-========================================
-
-.. toctree::
-   :maxdepth: 1
-
-   notebooks/ex0_input_data_description.ipynb
-   notebooks/ex1_EMGMeasurement.ipynb
-   notebooks/ex2_EMGMeasurementCollection.ipynb
-   notebooks/ex3_DataProcessingManager.ipynb
-   notebooks/ex4_DataProcessingManager.ipynb
