+++ docs/_sources/api/pyemgpipeline.processors.rst.txt
@@ -1,92 +1,0 @@
-﻿pyemgpipeline.processors
-========================================
-
-.. py:currentmodule:: pyemgpipeline.processors
-
-Summary
-----------------------------------------
-
-Base Class
-++++++++++++++++++++
-
-.. autosummary::
-   :nosignatures:
-
-   BaseProcessor
-
-Classes
-++++++++++++++++++++
-
-.. autosummary::
-   :nosignatures:
-
-   AmplitudeNormalizer
-   BandpassFilter
-   DCOffsetRemover
-   EndFrameCutter
-   FullWaveRectifier
-   LinearEnvelope
-   Segmenter
-
-AmplitudeNormalizer
-----------------------------------------
-
-.. autoclass:: AmplitudeNormalizer
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-BandpassFilter
-----------------------------------------
-
-.. autoclass:: BandpassFilter
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-BaseProcessor
-----------------------------------------
-
-.. autoclass:: BaseProcessor
-   :members:
-   :undoc-members:
-
-DCOffsetRemover
-----------------------------------------
-
-.. autoclass:: DCOffsetRemover
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-EndFrameCutter
-----------------------------------------
-
-.. autoclass:: EndFrameCutter
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-FullWaveRectifier
-----------------------------------------
-
-.. autoclass:: FullWaveRectifier
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-LinearEnvelope
-----------------------------------------
-
-.. autoclass:: LinearEnvelope
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Segmenter
-----------------------------------------
-
-.. autoclass:: Segmenter
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
