+++ docs/_sources/index.rst.txt
@@ -1,0 +1,26 @@
+pyemgpipeline Documentation
+========================================
+
+**pyemgpipeline** is an electromyography (EMG) signal processing pipeline package.
+It implements internationally accepted EMG processing conventions
+and provides a high-level interface for ensuring user adherence to those conventions,
+in terms of (1) processing parameter values, (2) processing steps, and (3) processing step order.
+It is suitable for both surface EMG and intramuscular EMG.
+
+Within this package, class :ref:`DataProcessingManager` is
+designed for high-level, guided processing,
+and users are encouraged to use it to adhere to accepted EMG processing conventions.
+
+The source code is available in the
+`GitHub repository <https://github.com/aalhossary/pyemgpipeline>`_.
+
+Contents
+----------------------------------------
+
+.. toctree::
+   :maxdepth: 2
+
+   installation.rst
+   quickstart.rst
+   examples.rst
+   api.rst

[…]
