### 2022-04-07 #4156 @tlwu
> MISSING DOIs
>
> - 10.1201/9781420036985.ax2 may be a valid DOI for title: Standards for reporting EMG data

This reference can be found by googling this DOI, but it cannot be found by searching the DOI at https://www.doi.org/. We've added this DOI in the paper.
