### 2022-04-07 #4156 @tlwu
>     * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.

The software is released. The version tag is v1.0.0_RC.

>     * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
>     * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>     * [ ]  Please list the DOI of the archived version here.

The software is archived in Zenodo, the metadata is checked, and the DOI of the archived version is 10.5281/zenodo.6422544.
