### 2022-04-08 aalhossary/pyemgpipeline#3 @osorensen
Thanks!
